Add PoC FreeBSD 11.1 hwpmc crash
When hwpmc is loaded/compiled into the kernel, running this will prevent
the kernel from forking any new processes requiring a hard reset of the
machine.

If security.bsd.unprivileged_proc_debug is 0 (the default) this can be
triggered by a unprivileged process.

Tested on FreeBSD 11.1-RELEASE-p8 running on amd64 with hwpmc loaded at
runtime but probably applies to other versions and architectures.